+++ src/interfaces/ISuperFormFactory.sol
@@ -41,7 +41,7 @@ interface ISuperFormFactory {
     /// @param salt_ is the salt for create2
     function addFormBeacon(
         address formImplementation_,
-        uint256 formBeaconId_,
+        uint32 formBeaconId_,
         bytes32 salt_
     ) external returns (address beacon);
 
@@ -51,7 +51,7 @@ interface ISuperFormFactory {
     /// @param salt_ is the salt for create2
     function addFormBeacons(
         address[] memory formImplementations_,
-        uint256[] memory formBeaconIds_,
+        uint32[] memory formBeaconIds_,
         bytes32 salt_
     ) external;
 
@@ -67,7 +67,7 @@ interface ISuperFormFactory {
     /// @return superFormId_ is the id of the superform
     /// @dev TODO: add array version of this
     function createSuperForm(
-        uint256 formBeaconId_,
+        uint32 formBeaconId_,
         address vault_,
         bytes calldata broadcastParams_
     ) external payable returns (uint256 superFormId_, address superForm_);
@@ -79,13 +79,13 @@ interface ISuperFormFactory {
     /// @dev allows an admin to update the logic of a form
     /// @param formBeaconId_ is the id of the form beacon
     /// @param newFormLogic_ is the address of the new form logic
-    function updateFormBeaconLogic(uint256 formBeaconId_, address newFormLogic_) external;
+    function updateFormBeaconLogic(uint32 formBeaconId_, address newFormLogic_) external;
 
     /// @dev allows an admin to change the status of a form
     /// @param formBeaconId_ is the id of the form beacon
     /// @param status_ is the new status
     /// @param extraData_ is optional & passed when broadcasting of status is needed
-    function changeFormBeaconPauseStatus(uint256 formBeaconId_, bool status_, bytes memory extraData_) external payable;
+    function changeFormBeaconPauseStatus(uint32 formBeaconId_, bool status_, bytes memory extraData_) external payable;
 
     /*///////////////////////////////////////////////////////////////
                             View Functions
@@ -94,12 +94,12 @@ interface ISuperFormFactory {
     /// @dev returns the address of a form beacon
     /// @param formBeaconId_ is the id of the beacon form
     /// @return formBeacon_ is the address of the beacon form
-    function getFormBeacon(uint256 formBeaconId_) external view returns (address formBeacon_);
+    function getFormBeacon(uint32 formBeaconId_) external view returns (address formBeacon_);
 
     /// @dev returns the status of form beacon
     /// @param formBeaconId_ is the id of the beacon form
     /// @return status_ is the current status of the form beacon
-    function getFormBeaconStatus(uint256 formBeaconId_) external view returns (bool status_);
+    function getFormBeaconStatus(uint32 formBeaconId_) external view returns (bool status_);
 
     /// @dev returns the address of a superform
     /// @param superFormId_ is the id of the superform
@@ -108,7 +108,7 @@ interface ISuperFormFactory {
     /// @return chainId_ is the chain id
     function getSuperForm(
         uint256 superFormId_
-    ) external pure returns (address superForm_, uint256 formBeaconId_, uint16 chainId_);
+    ) external pure returns (address superForm_, uint32 formBeaconId_, uint64 chainId_);
 
     /// @dev Reverse query of getSuperForm, returns all superforms for a given vault
     /// @param vault_ is the address of a vault
@@ -124,8 +124,8 @@ interface ISuperFormFactory {
         returns (
             uint256[] memory superFormIds_,
             address[] memory superForms_,
-            uint256[] memory formBeaconIds_,
-            uint16[] memory chainIds_
+            uint32[] memory formBeaconIds_,
+            uint64[] memory chainIds_
         );
 
     /// @dev Returns all SuperForms
@@ -139,8 +139,8 @@ interface ISuperFormFactory {
         returns (
             uint256[] memory superFormIds_,
             address[] memory vaults_,
-            uint256[] memory formBeaconIds_,
-            uint16[] memory chainIds_
+            uint32[] memory formBeaconIds_,
+            uint64[] memory chainIds_
         );
 
     /// @dev returns the number of forms
